[wrangler] Move r2/ commands to src/commands/r2/ (Phase 2c)
Move R2 storage commands into the src/commands/ directory structure.

Files moved (23 files):
- src/r2/*.ts → src/commands/r2/*.ts
- src/r2/helpers/*.ts → src/commands/r2/helpers/*.ts

Imports updated in:
- src/index.ts
- src/deployment-bundle/bindings.ts
- src/pipelines/client.ts
- src/__tests__/r2/*.ts

No logic changes - only file moves and import updates.
All 2827 tests pass.
