+++ packages/wrangler/src/commands/r2/catalog.ts
@@ -3,11 +3,11 @@ import {
 	getCloudflareApiEnvironmentFromEnv,
 	UserError,
 } from "@cloudflare/workers-utils";
-import { createCommand, createNamespace } from "../core/create-command";
-import { confirm } from "../dialogs";
-import { logger } from "../logger";
-import { requireAuth } from "../user";
-import formatLabelledValues from "../utils/render-labelled-values";
+import { createCommand, createNamespace } from "../../core/create-command";
+import { confirm } from "../../dialogs";
+import { logger } from "../../logger";
+import { requireAuth } from "../../user";
+import formatLabelledValues from "../../utils/render-labelled-values";
 import {
 	disableR2Catalog,
 	disableR2CatalogCompaction,
