+++ packages/wrangler/src/commands/r2/helpers/local-uploads.ts
@@ -1,4 +1,4 @@
-import { fetchResult } from "../../cfetch";
+import { fetchResult } from "../../../cfetch";
 import type { ComplianceConfig } from "@cloudflare/workers-utils";
 
 export interface LocalUploadsConfig {
