+++ packages/wrangler/src/commands/r2/helpers/misc.ts
