+++ packages/wrangler/src/commands/r2/helpers/notification.ts
@@ -1,7 +1,7 @@
-import { fetchResult } from "../../cfetch";
-import { logger } from "../../logger";
-import { getQueue, getQueueById } from "../../queues/client";
-import type { ApiCredentials } from "../../user";
+import { fetchResult } from "../../../cfetch";
+import { logger } from "../../../logger";
+import { getQueue, getQueueById } from "../../../queues/client";
+import type { ApiCredentials } from "../../../user";
 import type { Config } from "@cloudflare/workers-utils";
 import type { HeadersInit } from "undici";
