+++ packages/wrangler/src/commands/r2/index.ts
@@ -1,4 +1,4 @@
-import { createNamespace } from "../core/create-command";
+import { createNamespace } from "../../core/create-command";
 
 export const r2Namespace = createNamespace({
 	metadata: {
