+++ packages/wrangler/src/commands/r2/lifecycle.ts
@@ -1,10 +1,10 @@
 import { readFileSync, UserError } from "@cloudflare/workers-utils";
-import { createCommand, createNamespace } from "../core/create-command";
-import { confirm, multiselect, prompt } from "../dialogs";
-import isInteractive from "../is-interactive";
-import { logger } from "../logger";
-import { requireAuth } from "../user";
-import formatLabelledValues from "../utils/render-labelled-values";
+import { createCommand, createNamespace } from "../../core/create-command";
+import { confirm, multiselect, prompt } from "../../dialogs";
+import isInteractive from "../../is-interactive";
+import { logger } from "../../logger";
+import { requireAuth } from "../../user";
+import formatLabelledValues from "../../utils/render-labelled-values";
 import {
 	getLifecycleRules,
 	putLifecycleRules,
