+++ packages/wrangler/src/commands/r2/local-uploads.ts
@@ -1,7 +1,7 @@
-import { createCommand, createNamespace } from "../core/create-command";
-import { confirm } from "../dialogs";
-import { logger } from "../logger";
-import { requireAuth } from "../user";
+import { createCommand, createNamespace } from "../../core/create-command";
+import { confirm } from "../../dialogs";
+import { logger } from "../../logger";
+import { requireAuth } from "../../user";
 import {
 	getR2LocalUploadsConfig,
 	setR2LocalUploadsConfig,
