+++ packages/wrangler/src/commands/r2/public-dev-url.ts
@@ -1,7 +1,7 @@
-import { createCommand, createNamespace } from "../core/create-command";
-import { confirm } from "../dialogs";
-import { logger } from "../logger";
-import { requireAuth } from "../user";
+import { createCommand, createNamespace } from "../../core/create-command";
+import { confirm } from "../../dialogs";
+import { logger } from "../../logger";
+import { requireAuth } from "../../user";
 import { getR2DevDomain, updateR2DevDomain } from "./helpers/domain";
 
 export const r2BucketDevUrlNamespace = createNamespace({
